[Tom] Move the generated Constants to:
* Make use on generated decode() methods in generated message decoders
* Generated required field and group sets using Constant numbers
