+++ artio-core/src/main/java/uk/co/real_logic/artio/engine/logger/PossDupFinder.java
@@ -16,7 +16,7 @@
 package uk.co.real_logic.artio.engine.logger;
 
 import uk.co.real_logic.artio.ValidationError;
-import uk.co.real_logic.artio.decoder.Constants;
+import uk.co.real_logic.artio.Constants;
 import uk.co.real_logic.artio.fields.AsciiFieldFlyweight;
 import uk.co.real_logic.artio.otf.MessageControl;
 import uk.co.real_logic.artio.otf.OtfMessageAcceptor;
