+++ artio-core/src/main/java/uk/co/real_logic/artio/session/Session.java
@@ -21,6 +21,7 @@ import org.agrona.Verify;
 import org.agrona.concurrent.EpochClock;
 import org.agrona.concurrent.status.AtomicCounter;
 import uk.co.real_logic.artio.CommonConfiguration;
+import uk.co.real_logic.artio.Constants;
 import uk.co.real_logic.artio.Pressure;
 import uk.co.real_logic.artio.builder.Encoder;
 import uk.co.real_logic.artio.builder.HeaderEncoder;
@@ -40,8 +41,8 @@ import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static uk.co.real_logic.artio.builder.Validation.CODEC_VALIDATION_DISABLED;
 import static uk.co.real_logic.artio.builder.Validation.CODEC_VALIDATION_ENABLED;
-import static uk.co.real_logic.artio.decoder.Constants.NEW_SEQ_NO;
-import static uk.co.real_logic.artio.decoder.Constants.VERSION_CHARS;
+import static uk.co.real_logic.artio.Constants.NEW_SEQ_NO;
+import static uk.co.real_logic.artio.Constants.VERSION_CHARS;
 import static uk.co.real_logic.artio.dictionary.generation.CodecUtil.MISSING_INT;
 import static uk.co.real_logic.artio.dictionary.generation.CodecUtil.MISSING_LONG;
 import static uk.co.real_logic.artio.fields.RejectReason.*;
