+++ artio-system-tests/src/test/java/uk/co/real_logic/artio/system_tests/SystemTestUtil.java
@@ -22,7 +22,7 @@ import org.hamcrest.Matcher;
 import uk.co.real_logic.artio.CommonConfiguration;
 import uk.co.real_logic.artio.Reply;
 import uk.co.real_logic.artio.builder.TestRequestEncoder;
-import uk.co.real_logic.artio.decoder.Constants;
+import uk.co.real_logic.artio.Constants;
 import uk.co.real_logic.artio.engine.EngineConfiguration;
 import uk.co.real_logic.artio.engine.FixEngine;
 import uk.co.real_logic.artio.engine.LowResourceEngineScheduler;
